requirements.txt
python-telegram-bot[job-queue]==21.6
asyncpg==0.29.0
python-dotenv==1.0.1
google-generativeai==0.8.3
vnstock==0.2.9.2.3
ta==0.11.0
httpx==0.27.0
beautifulsoup4==4.12.3
yfinance==0.2.40
feedparser==6.0.11
pandas==2.2.2
pytz==2024.1
